% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation.R
\name{operation}
\alias{operation}
\title{Perform some operation}
\usage{
operation(x, y, op)
}
\arguments{
\item{x}{The first input value. This should only take a single value.}

\item{y}{The second input value. Should have the same dimensionality as first.}

\item{op}{The arithmetic operation to be run on x and y.}
}
\description{
Given two input values, perform some arithmetic operation on them.
}
